+++ CHANGELOG.md
@@ -11,6 +11,7 @@ CHANGELOG
 - `intelmq.lib.utils`:
   - log: Do not raise an exception if logging to neither file nor syslog is requested.
   - logging StreamHandler: Colorize all warning and error messages red.
+  - logging FileHandler: Strip all shell colorizations from the messages (#1436).
 
 ### Development